refactor: deduplicate code across auth, drivers, menus, and framebuffer
Extract shared helpers to eliminate ~680 lines of duplicated code:

- Auth: datetime_to_unix_timestamp delegates to der::DateTime, shared
  parse_der_length in mod.rs, shared RTC reading in new time.rs module
  (removes copies from crypto.rs, revocation.rs, dbx_update.rs)
- AHCI: extract_ata_model() and CommandHeader::init_atapi() replace
  duplicated model extraction (2x) and ATAPI header setup (3x)
- xHCI: get_config_descriptor() replaces identical 26-line blocks in
  configure_mass_storage and configure_hid_keyboard
- Varstore: write_record_to_storage() replaces 3 near-identical ~40-line
  write functions with thin wrappers
- Coreboot: iterate_table_records() replaces duplicated table parsing in
  parse_coreboot_tables and parse_forward
- Framebuffer: render_glyph() in framebuffer_console.rs replaces 3 copies
  of the glyph rendering loop; fb_log.rs delegates to it
- Menus: new menu_common.rs provides shared KeyPress, read_key(),
  clear_screen(), draw_header(), and SerialWriter for both menu.rs and
  secure_boot_menu.rs
